Cast doubles to floats explicitly
Explicitly cast doubles to floats to avoid narrowing warnings.

+++ main.cpp
@@ -166,9 +166,9 @@ OpenGLES2Test::timerEvent(QTimerEvent *event)
 
             float fgvtx[] = {
                 x + border, y + border,
-                x + border + (width - 2 * border) * reading / 9.81, y + border,
+                x + border + (width - 2 * border) * reading / 9.81f, y + border,
                 x + border, y + height - border,
-                x + border + (width - 2 * border) * reading / 9.81, y + height - border,
+                x + border + (width - 2 * border) * reading / 9.81f, y + height - border,
             };
 
             prog->setUniformValue("color", 1.0, 1.0, 1.0, 1.0);
@@ -194,10 +194,10 @@ OpenGLES2Test::timerEvent(QTimerEvent *event)
     QPointF c = center + (QPointF(-offset, offset) * m);
     QPointF d = center + (QPointF(offset, offset) * m);
     float vertices[] = {
-        a.x(), a.y(),
-        b.x(), b.y(),
-        c.x(), c.y(),
-        d.x(), d.y(),
+        (float)a.x(), (float)a.y(),
+        (float)b.x(), (float)b.y(),
+        (float)c.x(), (float)c.y(),
+        (float)d.x(), (float)d.y(),
     };
 
     glVertexAttribPointer(vtxcoord_loc, 2, GL_FLOAT, GL_FALSE, 0, vertices);
@@ -212,10 +212,10 @@ OpenGLES2Test::timerEvent(QTimerEvent *event)
     prog->setUniformValue("color", 0.0, 1.0, 0.0, 1.0);
     foreach (QPointF point, touches.values()) {
         float vtx[] = {
-            point.x()-offset, point.y()-offset,
-            point.x()+offset, point.y()-offset,
-            point.x()-offset, point.y()+offset,
-            point.x()+offset, point.y()+offset,
+            (float)point.x()-offset, (float)point.y()-offset,
+            (float)point.x()+offset, (float)point.y()-offset,
+            (float)point.x()-offset, (float)point.y()+offset,
+            (float)point.x()+offset, (float)point.y()+offset,
         };
         glVertexAttribPointer(vtxcoord_loc, 2, GL_FLOAT, GL_FALSE, 0, vtx);
         glDrawArrays(GL_TRIANGLE_STRIP, 0, 4);
